requirements.txt
streamlit
pandas
requests
beautifulsoup4
feedparser
newspaper3k
lxml_html_clean
sumy
jieba